api/requirements.txt
flask_wtf
flask
Pillow